+++ src/main/java/com/tutoring_calendar/models/Event.java
@@ -1,12 +1,11 @@
 package com.tutoring_calendar.models;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.sql.Date;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -24,7 +23,7 @@ public class Event {
     @JoinColumn(name = "client_id", nullable = false)
     private Client client;
     @Column(name = "price")
-    private Integer price;
+    private BigDecimal price;
     @Column(name = "date")
     private LocalDate date;
     @Column(name = "start_time")
